+++ src/main/kotlin/com/yourssu/openssupot/domain/domain/organization/OrganizationService.kt
@@ -1,6 +1,7 @@
 package com.yourssu.openssupot.domain.domain.organization
 
 import com.yourssu.openssupot.domain.domain.file.FileProcessor
+import com.yourssu.openssupot.domain.domain.password.PasswordValidator
 import com.yourssu.openssupot.domain.support.security.password.PasswordEncoder
 import org.springframework.stereotype.Service
